requirements.txt
# Core ML & Data Science
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.0.0

# LLM Integration
anthropic>=0.34.0

# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
nfl-data-py>=0.3.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Jupyter for experiments
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

